+++ packages/controllers/src/snaps/SnapController.test.ts
@@ -30,7 +30,9 @@ import * as utils from './utils';
 const { getSnapSourceShasum } = utils;
 
 const workerCode = fs.readFileSync(
-  require.resolve('@metamask/execution-environments/dist/webworker.bundle.js'),
+  require.resolve(
+    '@metamask/execution-environments/dist/webpack/webworker/bundle.js',
+  ),
   'utf8',
 );
 
@@ -94,7 +96,7 @@ const getSnapControllerOptions = (
     terminateAllSnaps: jest.fn(),
     terminateSnap: jest.fn(),
     executeSnap: jest.fn(),
-    endowmentPermissionNames: [],
+    environmentEndowmentPermissions: [],
     getRpcMessageHandler: jest.fn(),
     removeAllPermissionsFor: jest.fn(),
     getPermissions: jest.fn(),
@@ -116,7 +118,7 @@ const getSnapControllerWithEESOptions = (
   opts?: Partial<SnapControllerWithEESConstructorParams>,
 ) => {
   return {
-    endowmentPermissionNames: [],
+    environmentEndowmentPermissions: [],
     removeAllPermissionsFor: jest.fn(),
     getPermissions: jest.fn(),
     requestPermissions: jest.fn(),
@@ -406,7 +408,7 @@ describe('SnapController', () => {
 
     const snapController = getSnapController(
       getSnapControllerOptions({
-        endowmentPermissionNames: ['endowment:foo'],
+        environmentEndowmentPermissions: ['endowment:foo'],
         executeSnap: executeSnapMock,
         messenger,
       }),
@@ -679,7 +681,7 @@ describe('SnapController', () => {
       }),
     ).rejects.toThrow(/request timed out/u);
 
-    expect(snapController.state.snaps[snap.id].status).toStrictEqual('stopped');
+    expect(snapController.state.snaps[snap.id].status).toStrictEqual('crashed');
     snapController.destroy();
   });
 
@@ -1060,7 +1062,7 @@ describe('SnapController', () => {
         id: 1,
       }),
     ).rejects.toThrow(/request timed out/u);
-    expect(snapController.state.snaps[snap.id].status).toStrictEqual('stopped');
+    expect(snapController.state.snaps[snap.id].status).toStrictEqual('crashed');
 
     snapController.destroy();
   });
@@ -1136,7 +1138,7 @@ describe('SnapController', () => {
         id: 1,
       }),
     ).rejects.toThrow(/request timed out/u);
-    expect(snapController.state.snaps[snap.id].status).toStrictEqual('stopped');
+    expect(snapController.state.snaps[snap.id].status).toStrictEqual('crashed');
 
     await snapController.removeSnap(snap.id);